requirements.txt
requests==2.18.4
Flask==1.0.2
python-dateutil=1.11.0